Add files via upload
With the idea of (STREET FIXER) which focuses on solving the problems like potholes, broken streetlights,
clogged drains etc.. either in villages or Town. The main aim of our project is , It’s designed to be user friendly
and accessible, making it easy for residents to help improve their city. And by this Project.